Inventory module

+++ backend/src/main/java/org/io/GreenGame/inventory/service/InventoryServiceImplementation.java
@@ -27,15 +27,23 @@ public class InventoryServiceImplementation implements InventoryService {
             id = new Random().nextLong();
         }
         while (inventoryRepository.checkInventoryIDInDatabase(id) != 0);
-        if(userRepository.checkIfIdIsInDatabase(userID)==0) {
+        System.out.println("user: "+ userRepository.checkIfIdIsInDatabase(12L));
+        if(userRepository.checkIfIdIsInDatabase(userID) == 0L) {
+            System.out.println("No user in database");
             return false;
         }
         else {
             List<Item> items = Arrays.asList(new Item[10]);
+            /*TODO: po zmianach w Inventory (z user_id na GreenGameUser) ustawić:
+             * Inventory inventory = new Inventory(id, userRepository.findUserByID(userID), items, 0.0);
+             */
             Inventory inventory = new Inventory(id, userID, items, 0.0);
             try {
                 inventoryRepository.save(inventory);
+                System.out.println("Inventory saved");
             } catch (Exception e) {
+                System.out.println("Exception while saving inventory");
+                e.printStackTrace();
                 return false;
             }
         }
@@ -44,19 +52,46 @@ public class InventoryServiceImplementation implements InventoryService {
 
     @Override
     public Boolean addItemToInventory(Long userID, Long itemID) {
-        Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
-        Item tempItem = itemRepository.findItemByID(itemID);
-        Long inventoryID = tempInventory.getId();
-        if(inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+
+        if(inventoryRepository.findInventoryByUserID(userID) == null && itemRepository.findItemByID(itemID) == null) {
             return false;
         }
         else {
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            Item tempItem = itemRepository.findItemByID(itemID);
+            Long inventoryID = tempInventory.getId();
+            if(inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+                return false;
+            }
             try {
-                    tempInventory.addItem(tempInventory.getNextFreeSlotIndex(), tempItem);
+                    tempInventory.addItem(tempItem);
                     inventoryRepository.save(tempInventory);
                 } catch (Exception e) {
                     return false;
                 }
+        }
+        return true;
+    }
+
+    @Override
+    public Boolean addItemToInventory(Long userID, Item item) {
+
+
+        if(inventoryRepository.findInventoryByUserID(userID) == null) {
+            return false;
+        }
+        else {
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            Long inventoryID = tempInventory.getId();
+            if(inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+                return false;
+            }
+            try {
+                tempInventory.addItem(item);
+                inventoryRepository.save(tempInventory);
+            } catch (Exception e) {
+                return false;
+            }
         }
         return true;
     }
@@ -81,14 +116,41 @@ public class InventoryServiceImplementation implements InventoryService {
 
     @Override
     public Boolean deleteItemFromInventory(Long userID, Long itemID) {
-        Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
-        Item tempItem = itemRepository.findItemByID(itemID);
-        Long inventoryID = tempInventory.getId();
-        if (inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+        if(inventoryRepository.findInventoryByUserID(userID) == null || itemRepository.findItemByID(itemID) == null) {
+            System.out.println("Inventory not found or no item with given ID");
             return false;
-        } else {
+        }
+        else {
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            Item tempItem = itemRepository.findItemByID(itemID);
+            Long inventoryID = tempInventory.getId();
+            if (inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+                return false;
+            }
             try {
                 tempInventory.deleteItem(tempItem);
+                inventoryRepository.save(tempInventory);
+            } catch (Exception e) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public Boolean deleteItemFromInventory(Long userID, Item item) {
+        if(inventoryRepository.findInventoryByUserID(userID) == null) {
+            System.out.println("Inventory not found");
+            return false;
+        }
+        else {
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            Long inventoryID = tempInventory.getId();
+            if(inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
+                return false;
+            }
+            try {
+                tempInventory.deleteItem(item);
                 inventoryRepository.save(tempInventory);
             } catch (Exception e) {
                 return false;
@@ -118,6 +180,7 @@ public class InventoryServiceImplementation implements InventoryService {
     @Override
     public Inventory getUserInventory(Long userID) {
         if(userRepository.checkIfIdIsInDatabase(userID) == 0) {
+            System.out.println("No user in database");
             return null;
         }
         else {
@@ -156,14 +219,24 @@ public class InventoryServiceImplementation implements InventoryService {
 
     @Override
     public Item getItemFromInventory(Long userID, Long itemID) {
-        Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
-        Item tempItem = itemRepository.findItemByID(itemID);
-        Long inventoryID = tempInventory.getId();
-        if(inventoryRepository.checkInventoryIDInDatabase(inventoryID) == 0) {
-            return null;
+        if(inventoryRepository.findInventoryByUserID(userID) == null) {
+            throw new NullPointerException("Inventory not found");
+        } else if (itemRepository.findItemByID(itemID) == null) {
+            throw new NullPointerException("Item not found");
+        } else {
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            return tempInventory.getItem(itemID);
+        }
+    }
+
+    @Override
+    public Item getItemFromInventory(Long userID, Item item) {
+        if(inventoryRepository.findInventoryByUserID(userID) == null) {
+            throw new NullPointerException("Inventory not found");
         }
         else {
-            return tempInventory.getItem(itemID);
+            Inventory tempInventory = inventoryRepository.findInventoryByUserID(userID);
+            return tempInventory.getItem(item.getId());
         }
     }
 }
